% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDPC1}
\alias{GDPC1}
\title{Real Gross Domestic Product}
\format{
An object of class \code{ts}. From 1949Qtr1 to 2019Qtr2.
}
\source{
\url{https://fred.stlouisfed.org/series/GDPC1}
}
\usage{
GDPC1
}
\description{
Real gross domestic product is the inflation adjusted value of
the goods and services produced by labor and property located in the United States.
}
\references{
U.S. Bureau of Economic Analysis,
Real Gross Domestic Product [GDPC1], retrieved from FRED,
Federal Reserve Bank of St. Louis;
https://fred.stlouisfed.org/series/GDPC1, September 11, 2019.
}
\keyword{datasets}
