% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDPC1_GDPPOT}
\alias{GDPC1_GDPPOT}
\title{100*(Real Gross Domestic Product-Real Potential
Gross Domestic Product)/Real Potential Gross Domestic Product}
\format{
An object of class \code{ts}. From 1949Qtr1 to 2019Qtr2.
}
\source{
\url{https://fred.stlouisfed.org/graph/?g=f1cZ}
}
\usage{
GDPC1_GDPPOT
}
\description{
100*(Real Gross Domestic Product-Real Potential
Gross Domestic Product)/Real Potential Gross Domestic Product
}
\keyword{datasets}
