% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDPPOT}
\alias{GDPPOT}
\title{Real Potential Gross Domestic Product}
\format{
An object of class \code{ts}. From 1949Qtr1 to 2019Qtr2.
}
\source{
\url{https://fred.stlouisfed.org/series/GDPC1}
}
\usage{
GDPPOT
}
\description{
Real potential GDP is the CBO’s estimate of the output the
economy would produce with a high rate of use of its capital
and labor resources. The data is adjusted to remove
the effects of inflation.
}
\references{
U.S. Congressional Budget Office, Real Potential
Gross Domestic Product [GDPPOT], retrieved from FRED, Federal
Reserve Bank of St. Louis;
https://fred.stlouisfed.org/series/GDPPOT, September 11, 2019.
}
\keyword{datasets}
